refactor: 로그인, 회원가입 Activity 제거
